requirements.txt
streamlit
geopandas
fpdf
opencv-python-headless
seaborn
matplotlib
pandas
numpy
shapely
Pillow